python_server/requirement.txt
Flask==2.3.3
Flask-Cors==3.1.1
python-dotenv==1.0.0
google-generativeai==1.0.0 
